Initial core archetypes prep
Move archetypes to core.archetypes
Move referencedata to core.referencedata
Move references to the above
Disable $opal startproject for now
