+++ opal/core/archetypes.py
@@ -1,0 +1,165 @@
+"""
+This module will contain the core archetype clinical data model for OPAL
+"""
+# Model types
+from opal.models import PatientSubrecord, EpisodeSubrecord
+# Reference Data
+from opal.models import Destination, Drug, Drugroute, Drugfreq, Condition
+
+class Demographics(PatientSubrecord):
+    _is_singleton = True
+    _icon = 'fa fa-user'
+
+    name             = models.CharField(max_length=255, blank=True)
+    hospital_number  = models.CharField(max_length=255, blank=True)
+    nhs_number       = models.CharField(max_length=255, blank=True, null=True)
+    date_of_birth    = models.DateField(null=True, blank=True)
+    country_of_birth = ForeignKeyOrFreeText(Destination)
+    ethnicity        = models.CharField(max_length=255, blank=True, null=True)
+    gender           = models.CharField(max_length=255, blank=True, null=True)
+
+    class Meta:
+        abstract = True
+
+
+class Location(EpisodeSubrecord):
+    _is_singleton = True
+    _icon = 'fa fa-map-marker'
+
+    category = models.CharField(max_length=255, blank=True)
+    hospital = models.CharField(max_length=255, blank=True)
+    ward = models.CharField(max_length=255, blank=True)
+    bed = models.CharField(max_length=255, blank=True)
+
+    class Meta:
+        abstract = True
+
+    def __unicode__(self):
+        demographics = self.episode.patient.demographics_set.get()
+        return u'Location for {0}({1}) {2} {3} {4} {5}'.format(
+            demographics.name,
+            demographics.hospital_number,
+            self.category,
+            self.hospital,
+            self.ward,
+            self.bed
+            )
+
+
+class Treatment(EpisodeSubrecord):
+    _sort = 'start_date'
+    _icon = 'fa fa-flask'
+    _modal = 'lg'
+
+    drug          = ForeignKeyOrFreeText(Drug)
+    dose          = models.CharField(max_length=255, blank=True)
+    route         = ForeignKeyOrFreeText(Drugroute)
+    start_date    = models.DateField(null=True, blank=True)
+    end_date      = models.DateField(null=True, blank=True)
+    frequency     = ForeignKeyOrFreeText(Drugfreq)
+
+    class Meta:
+        abstract = True
+
+
+class Allergies(PatientSubrecord):
+    _icon = 'fa fa-warning'
+
+    drug        = ForeignKeyOrFreeText(Drug)
+    provisional = models.BooleanField(default=False)
+    details     = models.CharField(max_length=255, blank=True)
+
+    class Meta:
+        abstract = True
+
+
+class Diagnosis(EpisodeSubrecord):
+    """
+    This is a working-diagnosis list, will often contain things that are
+    not technically diagnoses, but is for historical reasons, called diagnosis.
+    """
+    _title = 'Diagnosis / Issues'
+    _sort = 'date_of_diagnosis'
+    _icon = 'fa fa-stethoscope'
+
+    condition         = ForeignKeyOrFreeText(Condition)
+    provisional       = models.BooleanField(default=False)
+    details           = models.CharField(max_length=255, blank=True)
+    date_of_diagnosis = models.DateField(blank=True, null=True)
+
+    class Meta:
+        abstract = True
+
+    def __unicode__(self):
+        return u'Diagnosis for {0}: {1} - {2}'.format(
+            self.episode.patient.demographics_set.get().name,
+            self.condition,
+            self.date_of_diagnosis
+            )
+
+
+class PastMedicalHistory(EpisodeSubrecord):
+    _title = 'PMH'
+    _sort = 'year'
+    _icon = 'fa fa-history'
+
+    condition = ForeignKeyOrFreeText(Condition)
+    year      = models.CharField(max_length=4, blank=True)
+    details   = models.CharField(max_length=255, blank=True)
+
+    class Meta:
+        abstract = True
+
+
+class Investigation(EpisodeSubrecord):
+    _title = 'Investigations'
+    _sort = 'date_ordered'
+    _icon = 'fa fa-crosshairs'
+    _modal = 'lg'
+
+    test                  = models.CharField(max_length=255)
+    date_ordered          = models.DateField(null=True, blank=True)
+    details               = models.CharField(max_length=255, blank=True)
+    microscopy            = models.CharField(max_length=255, blank=True)
+    organism              = models.CharField(max_length=255, blank=True)
+    sensitive_antibiotics = models.CharField(max_length=255, blank=True)
+    resistant_antibiotics = models.CharField(max_length=255, blank=True)
+    result                = models.CharField(max_length=255, blank=True)
+    igm                   = models.CharField(max_length=20, blank=True)
+    igg                   = models.CharField(max_length=20, blank=True)
+    vca_igm               = models.CharField(max_length=20, blank=True)
+    vca_igg               = models.CharField(max_length=20, blank=True)
+    ebna_igg              = models.CharField(max_length=20, blank=True)
+    hbsag                 = models.CharField(max_length=20, blank=True)
+    anti_hbs              = models.CharField(max_length=20, blank=True)
+    anti_hbcore_igm       = models.CharField(max_length=20, blank=True)
+    anti_hbcore_igg       = models.CharField(max_length=20, blank=True)
+    rpr                   = models.CharField(max_length=20, blank=True)
+    tppa                  = models.CharField(max_length=20, blank=True)
+    viral_load            = models.CharField(max_length=20, blank=True)
+    parasitaemia          = models.CharField(max_length=20, blank=True)
+    hsv                   = models.CharField(max_length=20, blank=True)
+    vzv                   = models.CharField(max_length=20, blank=True)
+    syphilis              = models.CharField(max_length=20, blank=True)
+    c_difficile_antigen   = models.CharField(max_length=20, blank=True)
+    c_difficile_toxin     = models.CharField(max_length=20, blank=True)
+    species               = models.CharField(max_length=20, blank=True)
+    hsv_1                 = models.CharField(max_length=20, blank=True)
+    hsv_2                 = models.CharField(max_length=20, blank=True)
+    enterovirus           = models.CharField(max_length=20, blank=True)
+    cmv                   = models.CharField(max_length=20, blank=True)
+    ebv                   = models.CharField(max_length=20, blank=True)
+    influenza_a           = models.CharField(max_length=20, blank=True)
+    influenza_b           = models.CharField(max_length=20, blank=True)
+    parainfluenza         = models.CharField(max_length=20, blank=True)
+    metapneumovirus       = models.CharField(max_length=20, blank=True)
+    rsv                   = models.CharField(max_length=20, blank=True)
+    adenovirus            = models.CharField(max_length=20, blank=True)
+    norovirus             = models.CharField(max_length=20, blank=True)
+    rotavirus             = models.CharField(max_length=20, blank=True)
+    giardia               = models.CharField(max_length=20, blank=True)
+    entamoeba_histolytica = models.CharField(max_length=20, blank=True)
+    cryptosporidium       = models.CharField(max_length=20, blank=True)
+
+    class Meta:
+        abstract = True
