+++ opal/models.py
@@ -23,7 +23,7 @@ from django.utils import timezone
 from django.conf import settings
 import reversion
 
-from opal.core import application, exceptions, lookuplists, plugins
+from opal.core import application, exceptions, plugins
 from opal import managers
 from opal.utils import camelcase_to_underscore
 from opal.core.fields import ForeignKeyOrFreeText
@@ -984,374 +984,6 @@ class Tagging(TrackedModel, models.Model):
 
         historic = Episode.objects.filter(id__in=eids)
         return historic
-
-
-"""
-Base Lookup Lists
-"""
-
-class Antimicrobial_route(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Antimicrobial route"
-
-
-class Antimicrobial(lookuplists.LookupList): pass
-
-
-class Antimicrobial_adverse_event(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Antimicrobial adverse event"
-
-
-class Antimicrobial_frequency(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Antimicrobial frequency"
-        verbose_name_plural = "Antimicrobial frequencies"
-
-
-class Clinical_advice_reason_for_interaction(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Clinical advice reason for interaction"
-        verbose_name_plural = "Clinical advice reasons for interaction"
-
-class Condition(lookuplists.LookupList): pass
-class Destination(lookuplists.LookupList): pass
-class Drug(lookuplists.LookupList): pass
-
-
-class Drugfreq(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Drug frequency"
-        verbose_name_plural = "Drug frequencies "
-
-
-class Drugroute(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Drug route"
-
-
-class Duration(lookuplists.LookupList): pass
-
-
-class Ethnicity(lookuplists.LookupList):
-    class Meta:
-        verbose_name_plural = "Ethnicities"
-
-
-class Gender(lookuplists.LookupList): pass
-class Hospital(lookuplists.LookupList): pass
-class Ward(lookuplists.LookupList): pass
-class Speciality(lookuplists.LookupList): pass
-
-# These should probably get refactored into opal-opat in 0.5
-class Line_complication(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Line complication"
-
-
-class Line_removal_reason(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Line removal reason"
-
-
-class Line_site(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Line site"
-
-
-class Line_type(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Line type"
-
-
-class Micro_test_c_difficile(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test C difficile"
-        verbose_name_plural = "Micro tests C difficile"
-
-
-class Micro_test_csf_pcr(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test CSF PCR"
-        verbose_name_plural = "Micro tests CSF PCR"
-
-
-class Micro_test_ebv_serology(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test EBV serology"
-        verbose_name_plural = "Micro tests EBV serology"
-
-
-class Micro_test_hepititis_b_serology(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test hepatitis B serology"
-        verbose_name_plural = "Micro tests hepatitis B serology"
-
-
-class Micro_test_hiv(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test HIV"
-        verbose_name_plural = "Micro tests HIV"
-
-
-class Micro_test_leishmaniasis_pcr(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test leishmaniasis PCR"
-        verbose_name_plural = "Micro tests leishmaniasis PCR"
-
-
-class Micro_test_mcs(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test MCS"
-        verbose_name_plural = "Micro tests MCS"
-
-
-class Micro_test_other(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test other"
-        verbose_name_plural = "Micro tests other"
-
-class Micro_test_parasitaemia(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test parasitaemia"
-        verbose_name_plural = "Micro tests parasitaemia"
-
-
-class Micro_test_respiratory_virus_pcr(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test respiratory virus PCR"
-        verbose_name_plural = "Micro tests respiratory virus PCR"
-
-
-class Micro_test_serology(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test serology"
-        verbose_name_plural = "Micro tests serology"
-
-
-class Micro_test_single_igg_test(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test single IgG test"
-        verbose_name_plural = "Micro tests single IgG test"
-
-
-class Micro_test_single_test_pos_neg(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test single test pos neg"
-        verbose_name_plural = "Micro tests single test pos neg"
-
-
-class Micro_test_single_test_pos_neg_equiv(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test single test pos neg equiv"
-        verbose_name_plural = "Micro tests single test pos neg equiv"
-
-
-class Micro_test_stool_parasitology_pcr(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test stool parasitology PCR"
-        verbose_name_plural = "Micro tests stool parasitology PCR"
-
-
-class Micro_test_stool_pcr(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test stool PCR"
-        verbose_name_plural = "Micro tests stool PCR"
-
-
-class Micro_test_swab_pcr(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test swab PCR"
-        verbose_name_plural = "Micro tests swab PCR"
-
-
-class Micro_test_syphilis_serology(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test syphilis serology"
-        verbose_name_plural = "Micro tests syphilis serology"
-
-
-class Micro_test_viral_load(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Micro test viral load"
-        verbose_name_plural = "Micro tests viral load"
-
-
-class Microbiology_organism(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Microbiology organism"
-
-
-class Symptom(lookuplists.LookupList): pass
-
-
-class Travel_reason(lookuplists.LookupList):
-    class Meta:
-        verbose_name = "Travel reason"
-
-
-"""
-Base models
-"""
-
-
-class Demographics(PatientSubrecord):
-    _is_singleton = True
-    _icon = 'fa fa-user'
-
-    name             = models.CharField(max_length=255, blank=True)
-    hospital_number  = models.CharField(max_length=255, blank=True)
-    nhs_number       = models.CharField(max_length=255, blank=True, null=True)
-    date_of_birth    = models.DateField(null=True, blank=True)
-    country_of_birth = ForeignKeyOrFreeText(Destination)
-    ethnicity        = models.CharField(max_length=255, blank=True, null=True)
-    gender           = models.CharField(max_length=255, blank=True, null=True)
-
-    class Meta:
-        abstract = True
-
-
-class Location(EpisodeSubrecord):
-    _is_singleton = True
-    _icon = 'fa fa-map-marker'
-
-    category = models.CharField(max_length=255, blank=True)
-    hospital = models.CharField(max_length=255, blank=True)
-    ward = models.CharField(max_length=255, blank=True)
-    bed = models.CharField(max_length=255, blank=True)
-
-    class Meta:
-        abstract = True
-
-    def __unicode__(self):
-        demographics = self.episode.patient.demographics_set.get()
-        return u'Location for {0}({1}) {2} {3} {4} {5}'.format(
-            demographics.name,
-            demographics.hospital_number,
-            self.category,
-            self.hospital,
-            self.ward,
-            self.bed
-            )
-
-
-class Treatment(EpisodeSubrecord):
-    _sort = 'start_date'
-    _icon = 'fa fa-flask'
-    _modal = 'lg'
-
-    drug          = ForeignKeyOrFreeText(Drug)
-    dose          = models.CharField(max_length=255, blank=True)
-    route         = ForeignKeyOrFreeText(Drugroute)
-    start_date    = models.DateField(null=True, blank=True)
-    end_date      = models.DateField(null=True, blank=True)
-    frequency     = ForeignKeyOrFreeText(Drugfreq)
-
-    class Meta:
-        abstract = True
-
-
-class Allergies(PatientSubrecord):
-    _icon = 'fa fa-warning'
-
-    drug        = ForeignKeyOrFreeText(Drug)
-    provisional = models.BooleanField(default=False)
-    details     = models.CharField(max_length=255, blank=True)
-
-    class Meta:
-        abstract = True
-
-
-class Diagnosis(EpisodeSubrecord):
-    """
-    This is a working-diagnosis list, will often contain things that are
-    not technically diagnoses, but is for historical reasons, called diagnosis.
-    """
-    _title = 'Diagnosis / Issues'
-    _sort = 'date_of_diagnosis'
-    _icon = 'fa fa-stethoscope'
-
-    condition         = ForeignKeyOrFreeText(Condition)
-    provisional       = models.BooleanField(default=False)
-    details           = models.CharField(max_length=255, blank=True)
-    date_of_diagnosis = models.DateField(blank=True, null=True)
-
-    class Meta:
-        abstract = True
-
-    def __unicode__(self):
-        return u'Diagnosis for {0}: {1} - {2}'.format(
-            self.episode.patient.demographics_set.get().name,
-            self.condition,
-            self.date_of_diagnosis
-            )
-
-
-class PastMedicalHistory(EpisodeSubrecord):
-    _title = 'PMH'
-    _sort = 'year'
-    _icon = 'fa fa-history'
-
-    condition = ForeignKeyOrFreeText(Condition)
-    year      = models.CharField(max_length=4, blank=True)
-    details   = models.CharField(max_length=255, blank=True)
-
-    class Meta:
-        abstract = True
-
-
-class Investigation(EpisodeSubrecord):
-    _title = 'Investigations'
-    _sort = 'date_ordered'
-    _icon = 'fa fa-crosshairs'
-    _modal = 'lg'
-
-    test                  = models.CharField(max_length=255)
-    date_ordered          = models.DateField(null=True, blank=True)
-    details               = models.CharField(max_length=255, blank=True)
-    microscopy            = models.CharField(max_length=255, blank=True)
-    organism              = models.CharField(max_length=255, blank=True)
-    sensitive_antibiotics = models.CharField(max_length=255, blank=True)
-    resistant_antibiotics = models.CharField(max_length=255, blank=True)
-    result                = models.CharField(max_length=255, blank=True)
-    igm                   = models.CharField(max_length=20, blank=True)
-    igg                   = models.CharField(max_length=20, blank=True)
-    vca_igm               = models.CharField(max_length=20, blank=True)
-    vca_igg               = models.CharField(max_length=20, blank=True)
-    ebna_igg              = models.CharField(max_length=20, blank=True)
-    hbsag                 = models.CharField(max_length=20, blank=True)
-    anti_hbs              = models.CharField(max_length=20, blank=True)
-    anti_hbcore_igm       = models.CharField(max_length=20, blank=True)
-    anti_hbcore_igg       = models.CharField(max_length=20, blank=True)
-    rpr                   = models.CharField(max_length=20, blank=True)
-    tppa                  = models.CharField(max_length=20, blank=True)
-    viral_load            = models.CharField(max_length=20, blank=True)
-    parasitaemia          = models.CharField(max_length=20, blank=True)
-    hsv                   = models.CharField(max_length=20, blank=True)
-    vzv                   = models.CharField(max_length=20, blank=True)
-    syphilis              = models.CharField(max_length=20, blank=True)
-    c_difficile_antigen   = models.CharField(max_length=20, blank=True)
-    c_difficile_toxin     = models.CharField(max_length=20, blank=True)
-    species               = models.CharField(max_length=20, blank=True)
-    hsv_1                 = models.CharField(max_length=20, blank=True)
-    hsv_2                 = models.CharField(max_length=20, blank=True)
-    enterovirus           = models.CharField(max_length=20, blank=True)
-    cmv                   = models.CharField(max_length=20, blank=True)
-    ebv                   = models.CharField(max_length=20, blank=True)
-    influenza_a           = models.CharField(max_length=20, blank=True)
-    influenza_b           = models.CharField(max_length=20, blank=True)
-    parainfluenza         = models.CharField(max_length=20, blank=True)
-    metapneumovirus       = models.CharField(max_length=20, blank=True)
-    rsv                   = models.CharField(max_length=20, blank=True)
-    adenovirus            = models.CharField(max_length=20, blank=True)
-    norovirus             = models.CharField(max_length=20, blank=True)
-    rotavirus             = models.CharField(max_length=20, blank=True)
-    giardia               = models.CharField(max_length=20, blank=True)
-    entamoeba_histolytica = models.CharField(max_length=20, blank=True)
-    cryptosporidium       = models.CharField(max_length=20, blank=True)
-
-    class Meta:
-        abstract = True
 
 
 class Role(models.Model):
